Added and tested support for GalleryList object.  I had originally planned for the GalleryList to do all kinds of complicated weighting and balancing so that when you make a change it affects the minimum number of rows possible.
But that was very complicated, so for now we're just serializing it and
storing the entire serialized array in the database.  Since we always
load the entire list, this isn't really that inefficient.  It'll eventually
be inefficient if we find ourselves in the situation where we're making
tiny changes to large albums.  At that point we'd want to adjust the way
the list works and have it page through the data instead of loading it all
at once.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@1754 57fcd75e-5312-0410-8df3-f5eb6fbb1595