Completely changed the way that SAVE and DELETE operations work. Because objects can be interconnected in all kinds of ways, it's pretty much impossible for the application layer to keep a handle on all the objects that will need to be saved to the database.
So, instead we now use the Gallery nucleus cache as a central
repository for items.  All load() calls draw items from this
cache, and save() calls put them back into the cache.  This
allows code in the class layer to retrieve an object, change
and save it, without interfering with other changes being made
by other parts of the class layer or the application layer.

This should be mostly transparent to the application layer.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@1805 57fcd75e-5312-0410-8df3-f5eb6fbb1595