Changed the way that permissions work.  We used to use BIT_OR() to aggregate permission data and then use a HAVING clause to pare down the result list to only send back the proper rows.  However, this doesn't work with row counts because the COUNT() aggregator happens before the HAVING filter so the results are incorrect.
So we now avoid this by getting rid of the ability to look for two
permissions at once.  In the case of search this means that we now
have a new comment.search composite permission that contains
(core.view, comment.view).  Both of these bits must be set in *one
permission row* in order for it to apply (both when you look for the
permission and when you look at it in the edit permission dialog).
This reduces some of our flexibility, but it's a more realistic
scope for permissions.

In the process, I completely redesigned the way that search works.
Now we let each module return localized key/value pairs to be displayed
on the search results page (instead of just returning item ids).  This
lets us show the actual comment result, instead of just the boilerplate
item values when you're searching the comment module.

Implemented tests for the core and comment search code (which is
how I found this problem in the first place).


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@2931 57fcd75e-5312-0410-8df3-f5eb6fbb1595