Changed the way that views return their results.  We now have the views return their results in an associative array so that we can allow a view to return a redirect.
This is very powerful, since now any view can (at any time) indicate that
it wants to redirect the user to a different page.  We use this in the
Matrix layout to allow us to navigate up and down the hierarchy and go
directly to the right page, without having to do page calculations
unnecessarily.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@4552 57fcd75e-5312-0410-8df3-f5eb6fbb1595