Changed toggle to be at the bottom and say "(Un)check all known types". This makes it more like ItemAddFromWeb's checkbox and reflects the fact that it's going to check/uncheck *known* types, not all boxes.
Change Javascript to work off of indices instead of using the name of
the element, which should make it more reliable in the case where the
name of the element is very weird (which it can be, since right now
the name of the element is the filename being uploaded).  This should
fix bug #1054407.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@8026 57fcd75e-5312-0410-8df3-f5eb6fbb1595