start session before outputting HTML in getImageBlock / search / searchScan
the reason why i don't add the $session->start() thing to ::init() if fullInit has been specified is that probably a lot of integrations always call fullInit and that would be bad in this case, also, it's just technically the wrong place.
so just remember to add $session->start(); $session->doNotUseTempId(); for all html ouputting exit points of g2.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@13170 57fcd75e-5312-0410-8df3-f5eb6fbb1595