Roll modules and themes together into AdminPlugins.  This gives us a unified way for administering themes.  The theme settings page now shows up in the SiteAdmin sidebar in the "Themes" group just like any other module, which makes for a better ui experience for the user.
Reworked how SiteAdmin handles the results from GalleryModule::getSiteAdminViews
a little bit so that we have a little more control over the site admin
view besides just the subView.  AdminTheme is still around to change the
default settings, and we still have a tabbed interface for modifying theme
settings (for now).


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@13485 57fcd75e-5312-0410-8df3-f5eb6fbb1595