Add backwards-compatible support for renderers.
Renderers provide the HTML output for an item without having to
be part of the item class itself.  This way we can associate
the panorama viewer with any GalleryPhotoItem (or any other type
that it could handle) without having to change the entity to
be a PanoramaPhotoItem or a PanoramaDerivativeImage.

* Introduce the new GalleryRenderer class
* Bump core module to 1.1.4
* Bump GalleryCoreApi version to 7.8
* Added code in both core and panorama module to revert
  PanoramaPhotoItem and PanoramaDerivativeImage back to
  their base types.  It's in both places to avoid issues
  where you upgrade core and not panorama, or vice versa.
* Added more tests for panorama module



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@13598 57fcd75e-5312-0410-8df3-f5eb6fbb1595