Allow users to convert their tables to UTF8 at any time.
Fixes bug 1561266.  If you upgrade from MySQL3 to MySQL4 we start
using UTF8, but your tables are still in whatever their old encoding
was.  With this change, we now try to do the table conversion at
upgrade time every time, and we provide a new maintenance task so that
you can do it by hand.

Bumps core module to 1.1.20



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@14897 57fcd75e-5312-0410-8df3-f5eb6fbb1595