Create a unified appearance for all of our support tools, give them a more pleasant style and consistent navigation.
Considerable rework to chmod.php.  Verified all of its code paths, and fixed a
bug where we were not properly allowing users to change the permissions back
and forth for individual plugins.  Removed references to /plugins/xxx since
DP didn't go in that direction.  Tightened up the text and removed stuff that
might be confusing to the user for no gain.

Verified on Unix with Firefox and IE7, but not yet on Windows.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@14919 57fcd75e-5312-0410-8df3-f5eb6fbb1595