It's literally 1 in a million (1 in 16^5) that the hash value for any two packages will collide.  But, if you work out the math, the odds are actually 50% that there will be a collision if you have 1206 packages:
  1/2 + sqrt(1/4 + 2 * (16^5) * l(2)) = 1206.168

We actually have something on the order of 4000+ packages now!  And it
turns out that panorama-lang-zh_TW-10212-10212.package collides with
thumbpage-lang-ja-10212-10212.package when we only look at the first 5
characters.  So let's bump it up to 10 characters which gives us:

  1/2 + sqrt(1/4 + 2 * (16^10) * l(2)) = 1234604.392

The odds on having a collision only hit 50% when we have 1.2M packages.
I think we'll be safe for a little while with this :-).

This won't affect existing packages.. it'll only get picked up when we
make new ones.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@15694 57fcd75e-5312-0410-8df3-f5eb6fbb1595