Fix for blocking bug #1620366.
Right now if you have any kind of platform failure on a DP download
(generally a mkdir or fopen failure) then you get platform errors
during the unpack phase followed by an integrity error.  With this
change we pre-verify all the package paths to make sure that they're
legal and are going to succeed before we do any unpacking.

It still may be possible to have an error, but we should figure out
every possible way it can fail and get that into the preverification
step so that we don't overwrite any files unless we're virtually
positive that it'll succeed.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@15743 57fcd75e-5312-0410-8df3-f5eb6fbb1595