Don't show "Upgrade All" button if there's nothing to upgrade
We don't actually check to see if there are any upgrades before
showing the button. We only used to check to see if there was any
metadata. But we take no useful action in the case where there's
nothing to be done (we don't even redirect after the post!). So don't
show the button in that case.

http://reviews.gallery2.org/index.jsp?page=ReviewDisplay&reviewid=87



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@15815 57fcd75e-5312-0410-8df3-f5eb6fbb1595