Guard DP against failed downloads.
Differentiate between packages that fail to download vs. ones that
fail to install. If we fail to download the base package, then don't
bother with the language updates either and don't reactivate the
package. If we fail to download a language, keep trying other
languages. Either way, show a list of files that failed to download.

Note: include some non-localized text here because we're in localization
      freeze for 2.2.  We will fix these ASAP in 2.3.

Review: http://reviews.gallery2.org/index.jsp?page=ReviewDisplay&reviewid=116



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@15930 57fcd75e-5312-0410-8df3-f5eb6fbb1595