Warn users when DP cannot automatically upgrade a plugin.
Some plugins may already be loaded (eg the default theme, or
authentication modules) so PHP will not let us evaluate the newly
downloaded code. This means that we can't upgrade those modules via
DP; the site admin must do it by hand on the AdminPlugins page. For
now, just notify the user.

Detecting this case requires us to use class_exists, so add that to
GalleryPhpVm to ease testing.

Added a utility method to the test code to extract the status from the
progressbar. There may be a much better way to do this, if not, then
this approach should probably be moved into GalleryTestCase in the
future. Also, the rest of the tests don't examine the status, and they
should.

Review #130: http://reviews.gallery2.org/index.jsp?page=ReviewDisplay&reviewid=130



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@15988 57fcd75e-5312-0410-8df3-f5eb6fbb1595