DP: Don't include the MANIFEST in bsae packages
We're not filtering the MANIFEST to be appropriate for what's actually
installed, so it includes localizations, unit test files, etc that are
not actually installed. So when we run through the integrity check at
upgrade time it bitches that you're missing files that you're supposed
to have.

The right solution is probably to manage the MANIFEST and add/remove
stuff as we add/remove packages. But the quick solution is to just
leave out the MANIFEST for now since we're doing an integrity check on
the module when we download and install it anyway.

In 2.3 we should bring MANIFEST files back and manage them properly.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@16013 57fcd75e-5312-0410-8df3-f5eb6fbb1595