DP: Activate plugins when they are first installed
My recent change which prevented activation of plugins that were
inactive broke the feature where we activate plugins when they are
first installed. This fixes it by checking to see if the plugin has a
version or not.

In order to do this, I had to intercept
GalleryCoreApi::fetchPluginList so I created a shim in the controller
that I can overload in the test code. Since I'm doing this, we can
stop calling $plugin->isActive and use the result from fetchPluginList
so I refactored that in the test.

The change to the controller code is relatively small. We should all
test it though with the following changes:

1. installing a plugin from scratch: it should get activated
2. upgrading an active plugin: it should stay active
3. upgrading an inactive plugin: it should stay inactive

Review #142.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@16018 57fcd75e-5312-0410-8df3-f5eb6fbb1595