Fix core upgrade notification in template
An incorrect variable name (which, sadly we don't get a warning for
from Smarty) causes us to fail to see the notification that there's a
newer version of the core available. And a quick review of the text
shows that one of the bullet points is confusing/unhelpful so I
removed it.

Review #173.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@16188 57fcd75e-5312-0410-8df3-f5eb6fbb1595