Say hello to the new slideshow!
I've replaced the old slideshow with two spiffy new ones. The main one
is PicLens, which is a slick, flashed-based slideshow backed by
CoolIris.com (a funded company). It can't be bundled with G2, so I've
added an admin interface to allow you to do a one-click
download/install of the SWF/JS files and a proxy view to send them
down to the client since they now live in the g2data dir.

If for some reason you can't or won't download PicLens, there's also a
new LyteBox based JS fallback slideshow which is still miles ahead of
the old slideshow.

This code requires the Gallery::BeforeDisplay event so that it can
inject a MediaRSS feed into the page which PicLens uses to make the
slideshow.



git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery2@17503 57fcd75e-5312-0410-8df3-f5eb6fbb1595